Upgrade VideoPlayer
The ConfigurePlayer page now supports loading and generating JSON data for runtime, in addition to defining and maintaining multiple overlays and multiple timelines per overlay. The VideoPlayer control now supports multiple buttons and multiple timelines per button, as well as the ability to turn on or off the highlight border and background color.