Geometry API updates for version 25.2711.4719
An error in casting from FVector2 to FPoint was suddenly causing exceptions to occur anywhere a function was returning an FVector2 by nature, but being used locally at the return site as an FPoint. This suggests that inheritance, as we know it no longer works - you are not even allowed to explicitly cast a base type to its extended type because it seems unsafe to someone. Without any further discussion on this topic, I am going to begin decoupling inherited objects and remaining as atomic at all times as possible, but for this version, I had implemented a work around where the return value is used as the construction parameter for creating a new FPoint as in *point = new FPoint(GetVertexValue(myVector2));*