Geometry API updates for version 25.2820.4317
A correction was made to the **FLine.Intersect(lineA, lineB, allowImaginary)** function, which was including estimations for points that had no possible intersections. That function now returns null if no intersection exists, the imaginary point, if imaginary is allowed, and null if an imaginary intersection exists but is not allowed.